Address PR #253 review comments
- Sort PublicAPI.Unshipped.txt so AndroidView entries land between
  AndroidView and AnimatedContent (previously inserted between
  AlertDialog.DismissButton and AlertDialog.Icon).
- AndroidView.cs: change broken `<see cref="Factory"/>` /
  `<see cref="Update"/>` to plain `<c>factory</c>` / `<c>update</c>`
  on the class-level remarks.
- Extract `AndroidViewFactoryAdapter` and `AndroidViewUpdateAdapter`
  into their own files (one-class-per-file convention; JCW
  `[Register]` requires non-nested types).
- LayoutHandler.cs: switch `_spacing` from `MutableState<int>` to
  `MutableState<float>` and store `(float)stack.Spacing`; route
  through `Arrangement.SpacedBy(new Dp(spacing))` to preserve
  fractional dp values.
- Rewrite `ComposeElementHandler<T>` description in
  `compose-maui.instructions.md` and `docs/maui-backend.md` — it is
  `ViewHandler<TVirtualView, ComposeView>`, not a marker
  `Android.Views.View`. Document the dual-mode behaviour: BuildNode
  for compose-aware parents (own `ComposeView` never attaches), own
  `ComposeView` as fallback for stock parents (CollectionView item
  templates, Border).
- Replace `BuildModifier()` (internal to Microsoft.AndroidX.Compose)
  with the public `Modifier` property pattern in the LayoutHandler
  example, matching real handler implementations.
- Update Phase 2 ButtonHandler skeleton to match the actual
  ButtonHandler.cs.
